refactor: reorganize file constants and add format filter options
